refactor: add schedule details, greatly simplify components related to schedule view
