+++ apps/food-planner-client/src/components/RecipePreview/types.ts
@@ -1,3 +1,3 @@
 export type TRecipePreviewProps = {
-  id: string;
+  id: string | number;
 };
