+++ apps/food-planner-client/src/components/ScheduleView/components/DailyMealsView/types.ts
@@ -1,7 +1,6 @@
-import type { EMealTypes, TScheduleMealsGet } from "@jcmono/api-contract";
+import type { TScheduleMealsGet } from "@jcmono/api-contract";
 
 export type DailyMealsViewProps = {
-  selectedDay: Date;
   meals: TScheduleMealsGet[];
-  onAddMeal: (day: Date, mealType: EMealTypes) => void;
+  currentViewDay: Date;
 };
