+++ apps/food-planner-client/src/components/ScheduleView/components/ScheduledMealCard/types.ts
@@ -1,0 +1,8 @@
+import type { TScheduleMealsGet } from "@jcmono/api-contract";
+
+export type ScheduledMealCardProps = {
+  foundMeal: TScheduleMealsGet;
+  handleDeleteMeal: (mealId: number) => void;
+  handleCreateShoppingList: (mealId: number) => void;
+  isCreateShoppingListFromMealPending: boolean;
+};
