+++ apps/food-planner-client/src/components/ScheduleView/components/SingleMealType/const.ts
