+++ apps/food-planner-client/src/components/ScheduleView/components/index.ts
@@ -1,5 +1,5 @@
 export { default as DailyMealsView } from "./DailyMealsView";
 export { default as DayNavigation } from "./DayNavigation";
-export { default as MealCard } from "./MealCard";
 export { default as ScheduleHeader } from "./ScheduleHeader";
+export { default as MealCard } from "./SingleMealType";
