+++ apps/food-planner-client/src/forms/ScheduleMealForm/types.ts
@@ -1,7 +1,6 @@
 import type { EMealTypes } from "@jcmono/api-contract";
 
 export type TScheduleMealFormProps = {
-  addMealDialogOpen: boolean;
   initialSelectedDay: Date | null;
   initialMealType: EMealTypes;
   handleCloseDialog: () => void;
